refactor(api): port import/epub + anki routes to Hono and proxy them
Last of the route ports:
- import/epub: parse a multipart EPUB upload (parseEpub / adm-zip, already
  in the api package) into a collection + lessons.
- anki: GET (version + deckNames) and POST (proxy an AnkiConnect action,
  auto-sync after addNote) against ANKI_CONNECT_URL.
- anki/sync-reviews: settings-aware AnkiConnect URL + 2.5s timeout; upsert
  per-day getNumCardsReviewedByDay into dailyStats.ankiReviews.

Registered in index.ts (auth SCOPE_MAP already covered import + anki).
Next routes become proxies. Verified through the proxy against a live
AnkiConnect (deck list + 188-day review sync) and a real EPUB upload
(multipart survives the proxy; 3 lessons created).
