+++ api/src/index.ts
@@ -12,6 +12,8 @@ import tts from './routes/tts';
 import extractUrl from './routes/extract-url';
 import dictionary from './routes/dictionary';
 import journal from './routes/journal';
+import importRoutes from './routes/import';
+import anki from './routes/anki';
 import lessons from './routes/lessons';
 import vocab from './routes/vocab';
 import cloze from './routes/cloze';
@@ -42,6 +44,8 @@ app.route('/api/tts', tts);
 app.route('/api/extract-url', extractUrl);
 app.route('/api/dictionary', dictionary);
 app.route('/api/journal', journal);
+app.route('/api/import', importRoutes);
+app.route('/api/anki', anki);
 app.route('/api/lessons', lessons);
 app.route('/api/vocab', vocab);
 app.route('/api/cloze', cloze);
